// src/app/control-panel/Models/customer.model.ts
export class Customer {
  id: number;
  fullname: string;
  email: string;
  phonecode: string;
  phonenumber: string;
  birthdate: string;
  cpf: string;
  rg: string;
  street: string;
  streetnumber: string;
  complement: string;
  district: string;
  city: string;
  state: string;
  country: string;
  zipcode: string;
  moip_id: string;
  created_at: string;
  updated_at: string;
}